Refactor basket operations and enhance Redis integration
- Simplified `GetAllUserIdAsync` in `ProductPriceChangedEventConsumer.cs` to call repository method directly.
- Added new methods in `BasketGrpcService.cs` for `UpdateBasket`, `Checkout`, and `DeleteBasket` operations.
- Changed `Price` type from `decimal` to `double` in `BasketItemDto`.
- Updated `ShoppingCart` constructor and `AddItem` method to use `UserId` and `double` for price.
- Introduced `GetAllUserIdsAsync` method in `IBasketRepository` and implemented it in `RedisBasketRepository.cs`.
- Added `Newtonsoft.Json` package reference in the project file for JSON serialization.
- Created `PrivateSetterContractResolver` to support private setters in JSON serialization.
- Defined HTTP request structure for creating a basket in `ECommerce.Basket.API.http`.
- Introduced `ShoppingCartDto` and `ShoppingCartItemDto` classes for data transfer between application and Redis.

+++ src/Services/Basket/ECommerce.Basket.Application/Features/Commands/CreateBasketCommandHandler.cs
@@ -10,7 +10,7 @@ using System.Threading.Tasks;
 
 namespace ECommerce.Basket.Application.Features.Commands
 {
-    public record BasketItemDto(int ProductId, string ProductName, string ProductImageUrl, decimal Price, int Quantity);
+    public record BasketItemDto(int ProductId, string ProductName, string ProductImageUrl, double Price, int Quantity);
     public record CreateBasketCommand(string UserId, string UserName, List<BasketItemDto> Items) : IRequest<Result<ShoppingCart>>;
     internal class CreateBasketCommandHandler : IRequestHandler<CreateBasketCommand, Result<ShoppingCart>>
     {
